Big progress new UI. Everything working good, audio window
