+++ mcu/src/mcu/databaseUtils.py
@@ -100,12 +100,12 @@ class ContentDatabase:
         # Run the initialization function
         initialisation_func(self)
 
-    def debug_log(self, message: str):
+    def debug_log(self, message: str) -> None:
         """ Log a message if debug is enabled (To reduce log spam, even in debug mode). """
         if self.debug:
             self.logger.debug(message)
 
-    def create_category(self, category_name: str):
+    def create_category(self, category_name: str) -> None:
         """ Create a new category in the database. """
         with self.lock:
             # Check if the category already exists
@@ -115,7 +115,7 @@ class ContentDatabase:
             else:
                 self.debug_log(f"Category already exists: {category_name}")
 
-    def add_item(self, category_name: str, item_name: str, item_value):
+    def add_item(self, category_name: str, item_name: str, item_value) -> None:
         """ Add a new item to a category in the database. """
         with self.lock:
             # Check if the category exists
@@ -129,7 +129,7 @@ class ContentDatabase:
             else:
                 self.debug_log(f"Category does not exist: {category_name}")
 
-    def get_item(self, category_name: str, item_name: str):
+    def get_item(self, category_name: str, item_name: str) -> "DatabaseElementTemplate":
         """ Get an item from a category in the database. """
         with self.lock:
             # Check if the category and item exist
@@ -143,20 +143,16 @@ class ContentDatabase:
                 self.debug_log(f"Category does not exist: {category_name}")
             return None
 
-    def gen_category_widget(self, category_name: str):
+    def gen_category_widget(self, category_name: str) -> QGroupBox:
         """ Generate a widget for a category in the database (using each element's generation function). """
         with self.lock:
             # Check if the category exists
             if category_name in self.db:
                 category_widget = QGroupBox(category_name)
                 category_layout = QVBoxLayout()
-                # Add items to the category
+                # Add all full widgets for each item in the category
                 for item_name, item_value in self.db[category_name].items():
-                    category_layout.addWidget(item_value.gen_widget())
-
-                # Test add the last widget twice
-                if len(self.db[category_name]) > 0:
-                    category_layout.addWidget(list(self.db[category_name].values())[-1].gen_widget())
+                    category_layout.addWidget(item_value.gen_widget_full())
 
                 # Push up the items
                 category_layout.addStretch()
@@ -165,14 +161,14 @@ class ContentDatabase:
             else:
                 self.debug_log(f"Category does not exist: {category_name}")
 
-    def reset_database(self):
+    def reset_database(self) -> None:
         """ Reset the database to an empty state, by signaling all items to reset. """
         with self.lock:
             for category in self.db.values():
                 for item in category.values():
                     item.trigger_callbacks()
 
-    def import_database(self, npy_file: pathlib.Path):
+    def import_database(self, npy_file: pathlib.Path) -> None:
         """ Import the values of the elements from a numpy file. """
         with self.lock:
             try:
@@ -188,7 +184,7 @@ class ContentDatabase:
             except Exception as e:
                 self.logger.error(f"Failed to import database: {e}")
 
-    def export_database(self, npy_file: pathlib.Path):
+    def export_database(self, npy_file: pathlib.Path) -> None:
         """ Export the values of the elements to a numpy file. """
         with self.lock:
             data = {}
@@ -204,11 +200,13 @@ class ContentDatabase:
             except Exception as e:
                 self.logger.error(f"Failed to export database: {e}")
 
+
+    # ------------------ Database Element Template ------------------
     class DatabaseElementTemplate:
         """
         Template for the database element
         """
-        def __init__(self, name: str, description: str, value): 
+        def __init__(self, name: str, value, description: str = ""): 
             self.lock = Lock()
             self.updating = False
             self.name = name
@@ -216,11 +214,14 @@ class ContentDatabase:
             self.value = value
             self.callbacks = []
 
-        def register_callback(self, callback: callable):
+        def register_callback(self, callback: callable) -> None:
+            """ Register a callback to be triggered when the value changes. """
             with self.lock:
                 # Add a callback to the list
                 self.callbacks.append(callback)
-        def trigger_callbacks(self):
+
+        def trigger_callbacks(self) -> None:
+            """ Trigger all registered callbacks. """
             with self.lock:
                 # Trigger all callbacks
                 self.updating = True
@@ -228,10 +229,12 @@ class ContentDatabase:
                     callback(self.value)
                 self.updating = False
 
-        def get_value(self): 
+        def get_value(self) -> any: 
+            """ Get the value of the element. """
             return self.value
         
-        def set_value(self, value): 
+        def set_value(self, value) -> None:
+            """ Set the value of the element and trigger all callbacks. """
             # If the callback is updating, return (to avoid infinite loops)
             if self.updating:
                 return
@@ -240,9 +243,26 @@ class ContentDatabase:
             self.trigger_callbacks()
 
         # Implement these methods in the derived classes
-        def __str__(self): ...
-        def __repr__(self): ...
-        def gen_widget(self): ...
+        def __str__(self) -> str: ...
+        def __repr__(self) -> str: ...
+        def gen_widget_element(self) -> QWidget: ...
+
+        # This method can be overridden in the derived classes if needed
+        def gen_widget_label(self) -> QLabel:
+            """ Generate a label widget for the element. """
+            label = QLabel(self.name)
+            label.setToolTip(self.description)
+            label.setFixedWidth(100)
+            return label
+        
+        def gen_widget_full(self) -> QWidget: 
+            """ Generate a full widget for the element. """
+            widget = QWidget()
+            layout = QHBoxLayout()
+            layout.addWidget(self.gen_widget_label())
+            layout.addWidget(self.gen_widget_element())
+            widget.setLayout(layout)
+            return widget
 
     # ------------------ Database Element Types ------------------
 
@@ -253,8 +273,8 @@ class ContentDatabase:
         Value composition :
             tuple<float, str> : (value, suffix)
         """
-        def __init__(self, name: str, description: str, value): 
-            super().__init__(name, description, value)
+        def __init__(self, name: str, value, description: str = ""):  
+            super().__init__(name, value, description)
 
         def __str__(self): 
             scaled_value, suffix_prefix = float_to_scaled_suffix(self.value[0])
@@ -290,14 +310,11 @@ class ContentDatabase:
             # or you can also parse suffix_str fully if you want to re-evaluate the unit.
             new_val = scaled_suffix_to_float(numeric_val, suffix_str)
             self.set_value((new_val, self.value[1]))
-
-        def gen_widget(self):
+        
+        def gen_widget_element(self) -> QWidget:
             scaled_value, suffix_prefix = float_to_scaled_suffix(self.value[0])
             widget = QWidget()
             layout = QHBoxLayout()
-            label = QLabel(self.name)
-            label.setToolTip(self.description)
-            label.setFixedWidth(100)
             text_entry = QLineEdit(str(scaled_value) + " " + suffix_prefix) # Editable text entry
             suffix_label = QLabel(self.value[1]) # Constant suffix (callback to update)
             text_entry.returnPressed.connect(lambda: self.handle_text_entry(text_entry))
@@ -306,7 +323,6 @@ class ContentDatabase:
                 scaled_value, suffix_prefix = float_to_scaled_suffix(value[0])
                 text_entry.setText(str(scaled_value) + " " + suffix_prefix)
             self.register_callback(handle_text_entry_callback)
-            layout.addWidget(label)
             layout.addWidget(text_entry)
             layout.addWidget(suffix_label)
             widget.setLayout(layout)
@@ -319,8 +335,8 @@ class ContentDatabase:
         Value composition :
             int : value
         """
-        def __init__(self, name: str, description: str, value): 
-            super().__init__(name, description, value)
+        def __init__(self, name: str, value, description: str = ""):  
+            super().__init__(name, value, description)
 
         def __str__(self): 
             return str(self.value)
@@ -328,22 +344,18 @@ class ContentDatabase:
         def __repr__(self): 
             return f"Integer: {self.name} - {self.value}"
         
-        def gen_widget(self):
+        def gen_widget_element(self) -> QWidget:
             widget = QWidget()
             layout = QHBoxLayout()
-            label = QLabel(self.name)
-            label.setToolTip(self.description)
-            label.setFixedWidth(100)
             spin_box = QSpinBox()
             spin_box.setRange(-2**31, 2**31 - 1)
             spin_box.setValue(self.value)
             spin_box.valueChanged.connect(lambda value: self.set_value(value))
             self.register_callback(lambda value: spin_box.setValue(value))
-            layout.addWidget(label)
             layout.addWidget(spin_box)
             widget.setLayout(layout)
             return widget
-        
+
     class RangeInt(DatabaseElementTemplate):
         """
         Integer range value
@@ -351,8 +363,8 @@ class ContentDatabase:
         Value composition :
             tuple<int, int, int> : (value, min, max)
         """
-        def __init__(self, name: str, description: str, value): 
-            super().__init__(name, description, value)
+        def __init__(self, name: str, value, description: str = ""):  
+            super().__init__(name, value, description)
 
         def __str__(self): 
             return f"{self.value[0]} ({self.value[1]}-{self.value[2]})"
@@ -360,13 +372,10 @@ class ContentDatabase:
         def __repr__(self): 
             return f"Range Integer: {self.name} - {self.value}"
         
-        def gen_widget(self):
+        def gen_widget_element(self) -> QWidget:
+            """ Generate a widget for the range integer. """
             widget = QWidget()
             layout = QHBoxLayout(widget)
-
-            label = QLabel(self.name)
-            label.setToolTip(self.description)
-            label.setFixedWidth(100)
 
             # Make the slider horizontal
             slider = QSlider(Qt.Orientation.Horizontal)
@@ -392,7 +401,6 @@ class ContentDatabase:
                 value_label.setText(str(int(val[0])))
             ))
 
-            layout.addWidget(label)
             layout.addWidget(slider)
             layout.addWidget(value_label)
             return widget
@@ -404,10 +412,11 @@ class ContentDatabase:
         Value composition:
             tuple<float, float, float> : (value, min, max)
         """
-        def __init__(self, name: str, description: str, value):
-            super().__init__(name, description, value)
-            # Decide how many decimals you want. For 2 decimals, factor=100
-            self.factor = 100  
+        # Decide how many decimals you want. For 2 decimals, factor=100
+        factor = 100  
+
+        def __init__(self, name: str, value, description: str = ""):  
+            super().__init__(name, value, description)
 
         def __str__(self):
             return f"{self.value[0]} ({self.value[1]}-{self.value[2]})"
@@ -423,13 +432,10 @@ class ContentDatabase:
             """ Convert a float to the slider’s integer scale. """
             return int(float_val * self.factor)
         
-        def gen_widget(self):
+        def gen_widget_element(self) -> QWidget:
+            """ Generate a widget for the range float. """
             widget = QWidget()
             layout = QHBoxLayout(widget)
-
-            label = QLabel(self.name)
-            label.setToolTip(self.description)
-            label.setFixedWidth(100)
 
             # Create a horizontal QSlider
             slider = QSlider(Qt.Orientation.Horizontal)
@@ -460,7 +466,6 @@ class ContentDatabase:
                 value_label.setText(f"{val[0]:.2f}")
             ))
 
-            layout.addWidget(label)
             layout.addWidget(slider)
             layout.addWidget(value_label)
             return widget
@@ -472,25 +477,22 @@ class ContentDatabase:
         Value composition:
             str : value
         """
-        def __init__(self, name: str, description: str, value): 
-            super().__init__(name, description, value)
+        def __init__(self, name: str, value, description: str = ""):  
+            super().__init__(name, value, description)
 
         def __str__(self): 
             return self.value
         
         def __repr__(self): 
             return f"Text: {self.name} - {self.value}"
-        
-        def gen_widget(self):
+
+        def gen_widget_element(self) -> QWidget:
+            """ Generate a widget for the text element. """
             widget = QWidget()
             layout = QHBoxLayout()
-            label = QLabel(self.name)
-            label.setToolTip(self.description)
-            label.setFixedWidth(100)
             text_entry = QLineEdit(self.value)
             text_entry.textChanged.connect(lambda text: self.set_value(text))
             self.register_callback(lambda text: text_entry.setText(text))
-            layout.addWidget(label)
             layout.addWidget(text_entry)
             widget.setLayout(layout)
             return widget
@@ -502,8 +504,8 @@ class ContentDatabase:
         Value composition:
             tuple<int, int, int> : (r, g, b)
         """
-        def __init__(self, name: str, description: str, value): 
-            super().__init__(name, description, value)
+        def __init__(self, name: str, value, description: str = ""):  
+            super().__init__(name, value, description)
 
         def __str__(self): 
             return f"({self.value[0]}, {self.value[1]}, {self.value[2]})"
@@ -511,12 +513,10 @@ class ContentDatabase:
         def __repr__(self): 
             return f"Color: {self.name} - {self.value}"
         
-        def gen_widget(self):
+        def gen_widget_element(self) -> QWidget:
+            """ Generate a widget for the color element. """
             widget = QWidget()
             layout = QHBoxLayout()
-            label = QLabel(self.name)
-            label.setToolTip(self.description)
-            label.setFixedWidth(100)
             color_button = QPushButton("Choose Color")
             color_button.setStyleSheet(
                 f"background-color: rgb({self.value[0]}, {self.value[1]}, {self.value[2]});"  # Set initial color
@@ -525,7 +525,6 @@ class ContentDatabase:
             self.register_callback(lambda color: color_button.setStyleSheet(
                 f"background-color: rgb({color[0]}, {color[1]}, {color[2]});"
             ))
-            layout.addWidget(label)
             layout.addWidget(color_button)
             widget.setLayout(layout)
             return widget
@@ -542,8 +541,8 @@ class ContentDatabase:
         Value composition:
             bool : value
         """
-        def __init__(self, name: str, description: str, value): 
-            super().__init__(name, description, value)
+        def __init__(self, name: str, value, description: str = ""):  
+            super().__init__(name, value, description)
 
         def __str__(self): 
             return str(self.value)
@@ -551,12 +550,10 @@ class ContentDatabase:
         def __repr__(self): 
             return f"Boolean: {self.name} - {self.value}"
         
-        def gen_widget(self):
+        def gen_widget_element(self) -> QWidget:
+            """ Generate a widget for the boolean element. """
             widget = QWidget()
             layout = QHBoxLayout()
-            label = QLabel(self.name)
-            label.setToolTip(self.description)
-            label.setFixedWidth(100)
 
             check_box = QCheckBox()
             check_box.setChecked(self.value)
@@ -566,7 +563,6 @@ class ContentDatabase:
 
             self.register_callback(lambda new_value: check_box.setChecked(new_value))
 
-            layout.addWidget(label)
             layout.addWidget(check_box)
             widget.setLayout(layout)
             return widget
@@ -578,8 +574,8 @@ class ContentDatabase:
         Value composition:
             list<str> : value
         """
-        def __init__(self, name: str, description: str, value): 
-            super().__init__(name, description, value)
+        def __init__(self, name: str, value, description: str = ""):  
+            super().__init__(name, value, description)
 
         def __str__(self): 
             return ", ".join(self.value)
@@ -587,14 +583,10 @@ class ContentDatabase:
         def __repr__(self): 
             return f"List: {self.name} - {self.value}"
         
-        def gen_widget(self):
+        def gen_widget_element(self) -> QWidget:
+            """ Generate a widget for the list element. """
             widget = QWidget()
             layout = QVBoxLayout(widget)
-
-            label = QLabel(self.name)
-            label.setToolTip(self.description)
-            label.setFixedWidth(100)
-            layout.addWidget(label)
 
             for i, text in enumerate(self.value):
                 widget_row = QWidget()
@@ -629,8 +621,8 @@ class ContentDatabase:
         Value composition:
             dict<str, str> : value
         """
-        def __init__(self, name: str, description: str, value): 
-            super().__init__(name, description, value)
+        def __init__(self, name: str, value, description: str = ""):  
+            super().__init__(name, value, description)
 
         def __str__(self): 
             return ", ".join([f"{key}: {val}" for key, val in self.value.items()])
@@ -638,14 +630,10 @@ class ContentDatabase:
         def __repr__(self): 
             return f"Dictionary: {self.name} - {self.value}"
         
-        def gen_widget(self):
+        def gen_widget_element(self) -> QWidget:
+            """ Generate a widget for the dictionary element. """
             widget = QWidget()
             layout = QVBoxLayout(widget)
-
-            label = QLabel(self.name)
-            label.setToolTip(self.description)
-            label.setFixedWidth(100)
-            layout.addWidget(label)
 
             for key, value in self.value.items():
                 widget_row = QWidget()
@@ -680,8 +668,8 @@ class ContentDatabase:
         Value composition:
             pathlib.Path : value
         """
-        def __init__(self, name: str, description: str, value): 
-            super().__init__(name, description, value)
+        def __init__(self, name: str, value, description: str = ""):  
+            super().__init__(name, value, description)
 
         def __str__(self): 
             return str(self.value)
@@ -689,18 +677,15 @@ class ContentDatabase:
         def __repr__(self): 
             return f"File: {self.name} - {self.value}"
         
-        def gen_widget(self):
+        def gen_widget_element(self) -> QWidget:
+            """ Generate a widget for the file element. """
             widget = QWidget()
             layout = QHBoxLayout()
-            label = QLabel(self.name)
-            label.setToolTip(self.description)
-            label.setFixedWidth(100)
             file_entry = QLineEdit(str(self.value))
             file_entry.setReadOnly(True)
             file_button = QPushButton("Choose File")
             file_button.clicked.connect(self.handle_file_dialog)
             self.register_callback(lambda value: file_entry.setText(str(value)))
-            layout.addWidget(label)
             layout.addWidget(file_entry)
             layout.addWidget(file_button)
             widget.setLayout(layout)
@@ -718,8 +703,8 @@ class ContentDatabase:
         Value composition:
             pathlib.Path : value
         """
-        def __init__(self, name: str, description: str, value): 
-            super().__init__(name, description, value)
+        def __init__(self, name: str, value, description: str = ""):  
+            super().__init__(name, value, description)
 
         def __str__(self): 
             return str(self.value)
@@ -727,18 +712,15 @@ class ContentDatabase:
         def __repr__(self): 
             return f"Folder: {self.name} - {self.value}"
         
-        def gen_widget(self):
+        def gen_widget_element(self) -> QWidget:
+            """ Generate a widget for the folder element. """
             widget = QWidget()
             layout = QHBoxLayout()
-            label = QLabel(self.name)
-            label.setToolTip(self.description)
-            label.setFixedWidth(100)
             folder_entry = QLineEdit(str(self.value))
             folder_entry.setReadOnly(True)
             folder_button = QPushButton("Choose Folder")
             folder_button.clicked.connect(self.handle_folder_dialog)
             self.register_callback(lambda value: folder_entry.setText(str(value)))
-            layout.addWidget(label)
             layout.addWidget(folder_entry)
             layout.addWidget(folder_button)
             widget.setLayout(layout)
@@ -756,8 +738,8 @@ class ContentDatabase:
         Value composition:
             tuple<int, list<str>> : (index, choices)
         """
-        def __init__(self, name: str, description: str, value): 
-            super().__init__(name, description, value)
+        def __init__(self, name: str, value, description: str = ""):  
+            super().__init__(name, value, description)
 
         def __str__(self): 
             return self.value[1][self.value[0]]
@@ -765,18 +747,19 @@ class ContentDatabase:
         def __repr__(self): 
             return f"ChoiceBox: {self.name} - {self.value}"
         
-        def gen_widget(self):
+        def gen_widget_element(self) -> QWidget:
+            """ Generate a widget for the choice box element. """
             widget = QWidget()
             layout = QHBoxLayout()
-            label = QLabel(self.name)
-            label.setToolTip(self.description)
-            label.setFixedWidth(100)
             choice_box = QComboBox()
             choice_box.addItems(self.value[1])
             choice_box.setCurrentIndex(self.value[0])
             choice_box.currentIndexChanged.connect(lambda index: self.set_value((index, self.value[1])))
-            self.register_callback(lambda value: choice_box.setCurrentIndex(value[0]))
-            layout.addWidget(label)
+            def choice_box_callback(index):
+                choice_box.clear()
+                choice_box.addItems(index[1])
+                choice_box.setCurrentIndex(index[0])
+            self.register_callback(choice_box_callback)
             layout.addWidget(choice_box)
             widget.setLayout(layout)
             return widget
@@ -788,8 +771,8 @@ class ContentDatabase:
         Value composition:
             str : value
         """
-        def __init__(self, name: str, description: str, value): 
-            super().__init__(name, description, value)
+        def __init__(self, name: str, value, description: str = ""):  
+            super().__init__(name, value, description)
 
         def __str__(self): 
             return self.value
@@ -797,15 +780,12 @@ class ContentDatabase:
         def __repr__(self): 
             return f"ConstantText: {self.name} - {self.value}"
         
-        def gen_widget(self):
+        def gen_widget_element(self) -> QWidget:
+            """ Generate a widget for the constant text element. """
             widget = QWidget()
             layout = QHBoxLayout()
-            label = QLabel(self.name)
-            label.setToolTip(self.description)
-            label.setFixedWidth(100)
             text_label = QLabel(self.value)
             self.register_callback(lambda value: text_label.setText(value)) # Update the label
-            layout.addWidget(label)
             layout.addWidget(text_label)
             widget.setLayout(layout)
             return widget
@@ -818,24 +798,19 @@ def init_db(db: ContentDatabase):
     and populates it with one SuffixFloat example item.
     """
     db.create_category("Audio Settings")
-    audio_gain_item = db.SuffixFloat(
-        name="Gain",
-        description="Controls the gain of the audio signal",
-        value=(1e3, "Hz")  # Example: 1000 Hz
-    )
-    db.add_item("Audio Settings", "gain", audio_gain_item)
-    db.add_item("Audio Settings", "gain2", db.Integer("Gain2", "Controls the gain of the audio signal", 1000))
-    db.add_item("Audio Settings", "gain3", db.RangeInt("Gain3", "Controls the gain of the audio signal", (1000, 0, 2000)))
-    db.add_item("Audio Settings", "gain4", db.RangeFloat("Gain4", "Controls the gain of the audio signal", (1000, 0, 2000)))
-    db.add_item("Audio Settings", "gain5", db.Text("Gain5", "Controls the gain of the audio signal", "1000"))
-    db.add_item("Audio Settings", "gain6", db.Color("Gain6", "Controls the gain of the audio signal", (255, 0, 0)))
-    db.add_item("Audio Settings", "gain7", db.Boolean("Gain7", "Controls the gain of the audio signal", True))
-    db.add_item("Audio Settings", "gain8", db.List("Gain8", "Controls the gain of the audio signal", ["1000", "2000", "3000"]))
-    db.add_item("Audio Settings", "gain9", db.Dictionary("Gain9", "Controls the gain of the audio signal", {"key1": "1000", "key2": "2000"}))
-    db.add_item("Audio Settings", "gain10", db.File("Gain10", "Controls the gain of the audio signal", db.app_path / "example.txt"))
-    db.add_item("Audio Settings", "gain11", db.Folder("Gain11", "Controls the gain of the audio signal", db.app_path))
-    db.add_item("Audio Settings", "gain12", db.ChoiceBox("Gain12", "Controls the gain of the audio signal", (0, ["1000", "2000", "3000"])))
-    db.add_item("Audio Settings", "gain13", db.ConstantText("Gain13", "Controls the gain of the audio signal", "1000"))
+    db.add_item("Audio Settings", "freq"  , db.SuffixFloat( "Freq " , (1e3, "Hz")                       , "Frequency Control"))
+    db.add_item("Audio Settings", "gain2" , db.Integer(     "Gain2 ", 1000                              , "First Gain Control"))
+    db.add_item("Audio Settings", "gain3" , db.RangeInt(    "Gain3 ", (1000, 0, 2000)                   , "Second Gain Control"))
+    db.add_item("Audio Settings", "gain4" , db.RangeFloat(  "Gain4 ", (1000, 0, 2000)                   , "Third Gain Control"))
+    db.add_item("Audio Settings", "gain5" , db.Text(        "Gain5 ", "1000"                            , "Text Gain Control"))
+    db.add_item("Audio Settings", "gain6" , db.Color(       "Gain6 ", (255, 0, 0)                       , "Color Gain Control"))
+    db.add_item("Audio Settings", "gain7" , db.Boolean(     "Gain7 ", True                              , "Boolean Gain Control"))
+    db.add_item("Audio Settings", "gain8" , db.List(        "Gain8 ", ["1000", "2000", "3000"]          , "List Gain Control"))
+    db.add_item("Audio Settings", "gain9" , db.Dictionary(  "Gain9 ", {"key1": "1000", "key2": "2000"}  , "Dictionary Gain Control"))
+    db.add_item("Audio Settings", "gain10", db.File(        "Gain10", db.app_path / "example.txt"       , "File Gain Control"))
+    db.add_item("Audio Settings", "gain11", db.Folder(      "Gain11", db.app_path                       , "Folder Gain Control"))
+    db.add_item("Audio Settings", "gain12", db.ChoiceBox(   "Gain12", (0, ["1000", "2000", "3000"])     , "ChoiceBox Gain Control"))
+    db.add_item("Audio Settings", "gain13", db.ConstantText("Gain13", "1000"                            , "ConstantText Gain Control"))
 
 class MainWindow(QMainWindow):
     def __init__(self, db: ContentDatabase):
@@ -858,13 +833,13 @@ class MainWindow(QMainWindow):
         text_entry = QLineEdit("1.0")
         text_entry.setReadOnly(True)
         def text_entry_callback(value):
-            suffix_float = db.get_item("Audio Settings", "gain").value[0]
+            suffix_float = db.get_item("Audio Settings", "freq").value[0]
             integer = db.get_item("Audio Settings", "gain2").value
             range_int = db.get_item("Audio Settings", "gain3").value[0]
             range_float = db.get_item("Audio Settings", "gain4").value[0]
             text_entry.setText(str(suffix_float * integer * range_int * range_float))
         # Register the callback to update the text entry at any change
-        db.get_item("Audio Settings", "gain").register_callback(text_entry_callback)
+        db.get_item("Audio Settings", "freq").register_callback(text_entry_callback)
         db.get_item("Audio Settings", "gain2").register_callback(text_entry_callback)
         db.get_item("Audio Settings", "gain3").register_callback(text_entry_callback)
         db.get_item("Audio Settings", "gain4").register_callback(text_entry_callback)
