+++ mcu/src/mcu/uart_logs.log
@@ -1,0 +1,80 @@
+[2025-01-03 22:34:02] INFO    : UI created
+[2025-01-03 23:48:35] ERROR   : Invalid port: -- No Ports Detected --
+[2025-01-03 23:48:47] ERROR   : Invalid port: -- No Ports Detected --
+[2025-01-03 23:49:08] SUCCESS : Connected to COM6 at 115200 baud
+[2025-01-03 23:49:10] DEBUG   : Stopping reader
+[2025-01-03 23:49:11] SUCCESS : Connected to COM6 at 115200 baud
+[2025-01-03 23:49:12] DEBUG   : Stopping reader
+[2025-01-03 23:50:01] SUCCESS : Connected to COM6 at 115200 baud
+[2025-01-03 23:50:02] DEBUG   : Stopping reader
+[2025-01-03 23:50:10] ERROR   : Invalid port: COM6
+[2025-01-03 23:50:14] ERROR   : Invalid port: COM6
+[2025-01-03 23:50:19] ERROR   : Invalid port: -- No Ports Detected --
+[2025-01-03 23:50:20] ERROR   : Invalid port: -- No Ports Detected --
+[2025-01-03 23:50:21] ERROR   : Invalid port: -- No Ports Detected --
+[2025-01-03 23:52:12] ERROR   : Invalid port: -- No Ports Detected --
+[2025-01-03 23:52:14] SUCCESS : Connected to COM6 at 115200 baud
+[2025-01-03 23:52:16] DEBUG   : Stopping reader
+[2025-01-03 23:52:16] DEBUG   : Cleaned up resources
+[2025-01-03 23:52:16] INFO    : Reader stopped
+[2025-01-03 23:52:24] SUCCESS : Connected to COM6 at 115200 baud
+[2025-01-03 23:52:34] SUCCESS : Connected to COM6 at 115200 baud
+[2025-01-03 23:52:34] DEBUG   : Stopping reader
+[2025-01-03 23:52:34] DEBUG   : Cleaned up resources
+[2025-01-03 23:52:34] INFO    : Reader stopped
+[2025-01-03 23:52:36] SUCCESS : Connected to COM6 at 115200 baud
+[2025-01-03 23:52:37] DEBUG   : Stopping reader
+[2025-01-03 23:52:37] DEBUG   : Cleaned up resources
+[2025-01-03 23:52:37] INFO    : Reader stopped
+[2025-01-03 23:52:42] SUCCESS : Connected to COM6 at 115200 baud
+[2025-01-03 23:52:44] ERROR   : Serial port disconnected unexpectedly
+[2025-01-03 23:52:44] DEBUG   : Stopping reader
+[2025-01-03 23:52:44] ERROR   : Serial port disconnected unexpectedly
+[2025-01-03 23:52:44] INFO    : Received 9 bytes
+[2025-01-03 23:52:44] INFO    : Reader stopped
+[2025-01-03 23:52:44] DEBUG   : Cleaned up resources
+[2025-01-03 23:55:13] SUCCESS : Connected to COM6 at 115200 baud
+[2025-01-03 23:55:15] DEBUG   : Stopping reader
+[2025-01-03 23:55:15] DEBUG   : Cleaned up resources
+[2025-01-03 23:55:15] INFO    : Reader stopped
+[2025-01-03 23:55:20] SUCCESS : Connected to COM6 at 115200 baud
+[2025-01-03 23:55:25] DEBUG   : Cleaned up resources
+[2025-01-03 23:56:41] SUCCESS : Connected to COM6 at 115200 baud
+[2025-01-03 23:56:43] DEBUG   : Stopping reader
+[2025-01-03 23:56:43] DEBUG   : Cleaned up resources
+[2025-01-03 23:56:43] INFO    : Reader stopped
+[2025-01-03 23:56:45] SUCCESS : Connected to COM6 at 115200 baud
+[2025-01-03 23:56:47] ERROR   : Serial port disconnected unexpectedly
+[2025-01-03 23:56:47] DEBUG   : Stopping reader
+[2025-01-03 23:56:47] INFO    : >>TERMINATE
+[2025-01-03 23:56:47] INFO    : Reader stopped
+[2025-01-03 23:56:47] DEBUG   : Cleaned up resources
+[2025-01-03 23:58:05] SUCCESS : Connected to COM6 at 115200 baud
+[2025-01-03 23:58:06] DEBUG   : Stopping reader
+[2025-01-03 23:58:06] DEBUG   : Cleaned up resources
+[2025-01-03 23:58:06] SUCCESS : Reader stopped
+[2025-01-03 23:58:10] SUCCESS : Connected to COM6 at 115200 baud
+[2025-01-03 23:59:03] SUCCESS : Connected to COM6 at 115200 baud
+[2025-01-03 23:59:03] DEBUG   : Stopping reader
+[2025-01-03 23:59:03] DEBUG   : Cleaned up resources
+[2025-01-03 23:59:03] WARNING : Reader stopped
+[2025-01-03 23:59:32] SUCCESS : Connected to COM6 at 115200 baud
+[2025-01-03 23:59:36] INFO    : >>Hello world
+[2025-01-03 23:59:36] INFO    : >>[S2LP] Init OK
+[2025-01-03 23:59:47] DEBUG   : Stopping reader
+[2025-01-03 23:59:47] DEBUG   : Cleaned up resources
+[2025-01-04 00:01:23] SUCCESS : Connected to COM6 at 115200 baud
+[2025-01-04 00:01:23] DEBUG   : Stopping reader
+[2025-01-04 00:01:23] DEBUG   : Cleaned up resources
+[2025-01-04 00:01:23] TRACE   : Reader stopped
+[2025-01-04 00:21:40] INFO    : Opening the Parameters Window
+[2025-01-04 00:26:20] INFO    : Opening the Parameters Window
+[2025-01-04 00:26:48] INFO    : Opening the Parameters Window
+[2025-01-04 00:28:14] INFO    : Opening the Parameters Window
+[2025-01-04 00:35:40] INFO    : Opening the Audio Window
+[2025-01-04 00:36:15] INFO    : Opening the Audio Window
+[2025-01-04 00:36:45] INFO    : Opening the Audio Window
+[2025-01-04 00:37:48] INFO    : Opening the Audio Window
+[2025-01-04 00:38:29] INFO    : Opening the Audio Window
+[2025-01-04 00:39:56] INFO    : Opening the Audio Window
+[2025-01-04 00:40:31] INFO    : Opening the Audio Window
